move base runner into runner directory
